update openssl : 1.1.1-pre9

+++ plesk-nginx.sh
@@ -133,8 +133,8 @@ fi
 distro_version=$(lsb_release -sc)
 
 if [[ "$NGINX_RELEASE" = "1" && "$RTMP" = "n" ]]; then
-    if [[ "$distro_version" == "xenial" || "$distro_version" == "bionic" ]]; then
-        if [[ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-bionic.list || ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-xenial.list ]]; then
+    if [ "$distro_version" == "xenial" ] || [ "$distro_version" == "bionic" ]; then
+        if [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-bionic.list ] && [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-xenial.list ]; then
             echo -ne "       Installing gcc-8                       [..]\\r"
             {
                 apt-get install software-properties-common -y
